+++ Engine/GameTemplate/GameTemplate/Game/Titan.h
@@ -7,7 +7,7 @@ public:
 	Titan();
 	~Titan() {};
 	void Attack();							//攻撃(今は未実装)
-	void Damage(float Damage) override;		//DAMAGE（今は未実装）
+	void Damage(int Damage) override;		//DAMAGE（今は未実装）
 	void Search();							//プレイヤーを見つける関数
 	void Update();							//敵の更新内容を記入
 	void Draw();							//エネミーをドロー処理だよ。
